refactor(ui): promote the System view to a top-level QAM page
Per-platform core selection + BIOS state move out of the library page sub-tab into a dedicated top-level System QAM page, reached from the main panel. LibraryPage keeps Platforms + Collections; index.tsx gains a system route and MainPage a System entry. The backend, the get_platform_core_info path, the get_firmware_status overview data source, and the active-core badge are unchanged.
