+++ docs/user-guide/bios-management.md
@@ -5,8 +5,8 @@ to launch. The plugin can download BIOS files directly from your RomM server.
 
 Which BIOS files a system needs depends on the **emulator core** in use — some cores need BIOS, some don't. Because the
 two concerns are related but independent, the plugin presents the **active core** and its **BIOS state** together in one
-place: the **System** tab. Core selection and BIOS file management can each be used on their own — they share a screen
-only because the active core determines which BIOS files matter.
+place: the **System** page, a top-level QAM destination. Core selection and BIOS file management can each be used on
+their own — they share a screen only because the active core determines which BIOS files matter.
 
 ## What Are BIOS Files?
 
@@ -37,12 +37,12 @@ Tap the BIOS status indicator to see a detailed list of individual files and whi
 
 ![BIOS file list overlay showing individual required files with checkmarks and "Missing" labels](../assets/screenshot-bios.jpg)
 
-## System Tab
+## System Page
 
-The **System** tab is the per-system emulator settings page: for each platform it shows the **active emulator core**
+The **System** page is the per-system emulator settings page: for each platform it shows the **active emulator core**
 first, then the BIOS files that core needs.
 
-1. From the main QAM page, open the library management page and tap **System**
+1. From the main QAM page, tap **System**
 2. Platforms with synced games that still need required BIOS files appear first, marked with "BIOS needed"
 3. For platforms with more than one available core, an **Emulator Core** dropdown is shown at the top of the platform's
    section — this is the primary per-system control
@@ -50,7 +50,7 @@ first, then the BIOS files that core needs.
 5. Tap **Show Files** to see the individual file list for a platform
 6. Tap **Download All** to download all missing BIOS files for a platform
 
-<!-- Screenshot: System tab showing per-platform Emulator Core dropdown above BIOS download counts -->
+<!-- Screenshot: System page showing per-platform Emulator Core dropdown above BIOS download counts -->
 
 BIOS files are downloaded to your RetroDECK bios directory (e.g. `~/retrodeck/bios/`). Some platforms use subdirectories
 — for example, Dreamcast BIOS goes into `bios/dc/` and PS2 BIOS goes into `bios/pcsx2/bios/`. The plugin handles the
@@ -80,7 +80,7 @@ your setup.
   replacement
 - With **gpSP**, `gba_bios.bin` is shown as _required_ — gpSP cannot run without it
 
-The active core name appears in both the game detail page (the **Emulator** column) and the **System** tab. This tells
+The active core name appears in both the game detail page (the **Emulator** column) and the **System** page. This tells
 you at a glance which core the plugin is filtering for.
 
 **How the core is determined:**
@@ -102,19 +102,19 @@ showing all files.
 You can change the active emulator core directly from the plugin, without leaving Game Mode. Changes are written to
 ES-DE's `gamelist.xml` so they persist across sessions and are picked up by both the plugin and ES-DE.
 
-### Per-Platform (System Tab)
+### Per-Platform (System Page)
 
-In the **System** tab, platforms with multiple available cores show an **Emulator Core** dropdown as the first control
+On the **System** page, platforms with multiple available cores show an **Emulator Core** dropdown as the first control
 in the platform's section, above the BIOS file list. Changing it sets the default core for all games on that platform. A
 "Switching cores may affect save compatibility" note appears under the dropdown for platforms that offer a choice.
 
-1. Open the library management page from the main QAM page and tap **System**
+1. Open the **System** page from the main QAM page
 2. Find the platform you want to change
 3. Use the **Emulator Core** dropdown to select a different core
 4. The BIOS file list below updates immediately to show files relevant to the new core
 
 This writes a system-wide override to ES-DE's `gamelist.xml`. ES-DE will pick up the change on next launch. The System
-tab works even when your RomM server is offline — core switching and BIOS status are available, only download buttons
+page works even when your RomM server is offline — core switching and BIOS status are available, only download buttons
 are disabled.
 
 ### Per-Game (Game Detail Page)
