+++ docs/user-guide/troubleshooting.md
@@ -10,7 +10,7 @@ Common issues and how to fix them.
 shows a black screen.
 
 **Fix**: Check the game's detail page — if the BIOS indicator is orange, you're missing required BIOS files. Go to the
-**System** tab in the QAM and tap **Download All** for the relevant platform. See [BIOS Management](bios-management.md)
+**System** page in the QAM and tap **Download All** for the relevant platform. See [BIOS Management](bios-management.md)
 for details.
 
 ### ROM not downloaded
@@ -161,7 +161,7 @@ games affected. This only removes local save files — saves already uploaded to
 ### Delete BIOS Files
 
 Deletes downloaded BIOS files for a specific platform. Shows each platform that has BIOS files present locally. You can
-re-download them later from the **System** tab.
+re-download them later from the **System** page.
 
 <!-- Screenshot: Danger Zone page showing removal options and whitelist -->
