+++ src/components/LibraryPage.test.tsx
@@ -1,9 +1,9 @@
 // CATCH-REJECTION ASSERTION RULE (applies to all orchestration shell tests):
 // Every catch block with a setX(...) side effect MUST have its side effect
-// asserted in the test (rolled-back toggle state, surfaced biosError /
-// biosStatus string, debugLog spy, etc.). Asserting only that the rejecting
-// call was invoked is vacuous — the rejection happens after the call returns
-// so the test would pass with or without the .catch.
+// asserted in the test (rolled-back toggle state, surfaced error string, etc.).
+// Asserting only that the rejecting call was invoked is vacuous — the rejection
+// happens after the call returns so the test would pass with or without the
+// .catch.
 //
 // LibraryPage catch sites (all asserted below):
 //   - handleToggle catch → rollback setSyncPlatforms (sync_enabled flips back)
@@ -11,20 +11,17 @@
 //   - handleCollectionToggle catch → rollback setCollections
 //   - handleSetAllCollections catch → restore previous collections snapshot
 //   - platform-groups inline catch → setPlatformGroups(!value) rollback
-//   - refreshBios try/catch → setBiosError(`Failed to fetch firmware status: ${e}`)
-//   - handleDownloadAll catch → setBiosStatus(`Download failed: ${e}`)
-//   - handleDownloadRequired catch → setBiosStatus(`Download failed: ${e}`)
 //   - getCollections/getSettings .catch → setCollectionsError(true)
+//
+// The System view (per-platform core + BIOS state) lives in SystemPage, a
+// top-level QAM page — its tests are in SystemPage.test.tsx, not here.
 //
 // MUTATION CHECKS (by inspection — auto-mode classifier likely blocks on
 // React state internals, so confidence is recorded here):
 //   1. Removing the rollback inside handleToggle's catch would break the
 //      "platform toggle rejection reverts checked state" test — the captured
 //      ToggleField checked prop would remain at the optimistic value.
-//   2. Removing setBiosError(result.message ...) from refreshBios's failure
-//      branch would break the "biosError surfaces result.message" test —
-//      the Error field's description would no longer render the message.
-//   3. Removing the `!collectionsLoaded.current` guard would break the
+//   2. Removing the `!collectionsLoaded.current` guard would break the
 //      "switching back to collections tab does not refetch" test — getCollections
 //      would be called twice instead of once.
 
@@ -32,25 +29,13 @@ import { describe, it, expect, beforeEach, vi } from "vitest";
 import { render, fireEvent, act } from "@testing-library/react";
 import { LibraryPage } from "./LibraryPage";
 import * as backend from "../api/backend";
-import type { PlatformSyncSetting, CollectionSyncSetting, FirmwarePlatformExt, PluginSettings } from "../types";
+import type { PlatformSyncSetting, CollectionSyncSetting, PluginSettings } from "../types";
 
 // scrollToTop is a no-op in happy-dom; mock for cleanliness.
 vi.mock("../utils/scrollHelpers", () => ({ scrollToTop: vi.fn() }));
 
-// DropdownItem in the global @decky/ui stub is a passthrough <select> with no
-// rgOptions / onChange capture. LibraryPage's BIOS tab uses DropdownItem for
-// per-platform core selection; we need to drive its onChange callback to test
-// the setSystemCore flow. Re-mock @decky/ui locally to expose DropdownItem
-// props on a shared captured-array — every other component mirrors the global
-// stub so the rest of the tree behaves identically.
-interface CapturedDropdown {
-  label?: string;
-  rgOptions?: Array<{ data: unknown; label: string }>;
-  selectedOption?: unknown;
-  onChange?: (option: { data: string; label: string }) => void | Promise<void>;
-}
-const capturedDropdowns: CapturedDropdown[] = [];
-
+// Re-mock @decky/ui locally so the component tree renders with inspectable
+// stubs (ToggleField checked-prop, Field label/description, DialogButton click).
 vi.mock("@decky/ui", async () => {
   const { createElement: ce } = await import("react");
   type AnyProps = Record<string, unknown> & { children?: unknown };
@@ -76,14 +61,6 @@ vi.mock("@decky/ui", async () => {
     Focusable: passthrough("div"),
     DialogButton: ({ children, onClick }: AnyProps & { onClick?: () => void }) =>
       ce("button", { onClick }, children as never),
-    DropdownItem: (p: CapturedDropdown) => {
-      capturedDropdowns.push(p);
-      return ce(
-        "div",
-        { "data-testid": "dropdown" },
-        ce("span", { "data-testid": "dropdown-label" }, p.label as never),
-      );
-    },
     ToggleField: (
       p: AnyProps & {
         checked?: boolean;
@@ -153,19 +130,9 @@ function makeCollection(overrides: Partial<CollectionSyncSetting> = {}): Collect
   };
 }
 
-function makeBiosPlatform(overrides: Partial<FirmwarePlatformExt> = {}): FirmwarePlatformExt {
-  return {
-    platform_slug: "snes",
-    files: [],
-    has_games: true,
-    ...overrides,
-  };
-}
-
 describe("LibraryPage", () => {
   beforeEach(() => {
     vi.resetAllMocks();
-    capturedDropdowns.length = 0;
     // Default callable behavior — tests override per case.
     vi.mocked(backend.getPlatforms).mockResolvedValue({
       success: true,
@@ -189,15 +156,6 @@ describe("LibraryPage", () => {
       success: true,
     });
     vi.mocked(backend.getSettings).mockResolvedValue(defaultSettings());
-    vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-      success: true,
-      platforms: [],
-    });
-    vi.mocked(backend.downloadAllFirmware).mockResolvedValue({ success: true });
-    vi.mocked(backend.downloadRequiredFirmware).mockResolvedValue({
-      success: true,
-    });
-    vi.mocked(backend.setSystemCore).mockResolvedValue({ success: true });
   });
 
   // ------------------------------------------------------------------
@@ -208,9 +166,8 @@ describe("LibraryPage", () => {
       render(<LibraryPage onBack={vi.fn()} />);
       await flushAsync();
       expect(vi.mocked(backend.getPlatforms)).toHaveBeenCalledTimes(1);
-      // collections + bios lazy data not yet fetched
+      // collections lazy data not yet fetched
       expect(vi.mocked(backend.getCollections)).not.toHaveBeenCalled();
-      expect(vi.mocked(backend.getFirmwareStatus)).not.toHaveBeenCalled();
     });
 
     it("clicking the Collections tab lazy-loads getCollections + getSettings", async () => {
@@ -223,17 +180,6 @@ describe("LibraryPage", () => {
       });
       expect(vi.mocked(backend.getCollections)).toHaveBeenCalledTimes(1);
       expect(vi.mocked(backend.getSettings)).toHaveBeenCalledTimes(1);
-    });
-
-    it("clicking the BIOS tab lazy-loads getFirmwareStatus", async () => {
-      const { getByText } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(vi.mocked(backend.getFirmwareStatus)).toHaveBeenCalledTimes(1);
     });
 
     it("switching back to Collections does NOT refetch (collectionsLoaded guard)", async () => {
@@ -255,26 +201,6 @@ describe("LibraryPage", () => {
       });
       // Still 1 — the ref guard prevents a re-fetch.
       expect(vi.mocked(backend.getCollections)).toHaveBeenCalledTimes(1);
-    });
-
-    it("switching back to BIOS does NOT refetch (biosLoaded guard)", async () => {
-      const { getByText } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(vi.mocked(backend.getFirmwareStatus)).toHaveBeenCalledTimes(1);
-      await act(async () => {
-        fireEvent.click(getByText("Platforms"));
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-      });
-      expect(vi.mocked(backend.getFirmwareStatus)).toHaveBeenCalledTimes(1);
     });
   });
 
@@ -1088,1007 +1014,6 @@ describe("LibraryPage", () => {
       } finally {
         warn.mockRestore();
       }
-    });
-  });
-
-  // ------------------------------------------------------------------
-  // I. BIOS tab — refreshBios
-  // ------------------------------------------------------------------
-  describe("bios tab — refreshBios", () => {
-    it("renders platforms and sets serverOffline on success", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        server_offline: false,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "snes.rom",
-                size: 100,
-                md5: "x",
-                downloaded: true,
-                required: true,
-                description: "BIOS",
-                hash_valid: true,
-                classification: "required",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("snes");
-    });
-
-    it("renders the server-offline banner when getFirmwareStatus reports server_offline", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        server_offline: true,
-        platforms: [],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("Server offline");
-    });
-
-    it("surfaces result.message when getFirmwareStatus returns success=false with a message", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: false,
-        message: "Server is sad",
-        platforms: [],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      // CATCH-REJECTION (failure branch): biosError = result.message
-      expect(container.textContent).toContain("Server is sad");
-    });
-
-    it("falls back to 'Failed to fetch firmware status' when result.message is absent", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: false,
-        platforms: [],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("Failed to fetch firmware status");
-    });
-
-    it("sets biosError='Failed to fetch firmware status: <e>' when getFirmwareStatus throws", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockRejectedValue(new Error("network"));
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      // CATCH-REJECTION assert: rendered with the interpolated Error
-      expect(container.textContent).toContain("Failed to fetch firmware status: Error: network");
-    });
-
-    it("renders the no-firmware empty state when platforms list is empty", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("No firmware files found");
-    });
-  });
-
-  // ------------------------------------------------------------------
-  // J. BIOS tab — handleDownloadAll
-  // ------------------------------------------------------------------
-  describe("bios tab — handleDownloadAll", () => {
-    function biosPlatformWithMissingOptional(): FirmwarePlatformExt {
-      return makeBiosPlatform({
-        platform_slug: "snes",
-        files: [
-          {
-            id: 1,
-            file_name: "boot.rom",
-            size: 100,
-            md5: "x",
-            downloaded: false,
-            required: false,
-            description: "Optional",
-            hash_valid: null,
-            classification: "optional",
-          },
-        ],
-      });
-    }
-
-    it("calls downloadAllFirmware(slug) and then refreshes BIOS on success", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [biosPlatformWithMissingOptional()],
-      });
-      vi.mocked(backend.downloadAllFirmware).mockResolvedValue({
-        success: true,
-        downloaded: 1,
-      });
-      const { getByText } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Download All"));
-        await Promise.resolve();
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(vi.mocked(backend.downloadAllFirmware)).toHaveBeenCalledWith("snes");
-      // refreshBios called once on tab activation + once after download
-      expect(vi.mocked(backend.getFirmwareStatus)).toHaveBeenCalledTimes(2);
-    });
-
-    it("surfaces result.message when the download succeeds", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [biosPlatformWithMissingOptional()],
-      });
-      vi.mocked(backend.downloadAllFirmware).mockResolvedValue({
-        success: true,
-        message: "All good",
-        downloaded: 1,
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Download All"));
-        await Promise.resolve();
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("All good");
-    });
-
-    it("surfaces 'Download failed' when result.success=false with no message", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [biosPlatformWithMissingOptional()],
-      });
-      vi.mocked(backend.downloadAllFirmware).mockResolvedValue({
-        success: false,
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Download All"));
-        await Promise.resolve();
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("Download failed");
-    });
-
-    it("sets biosStatus='Download failed: <e>' when downloadAllFirmware throws", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [biosPlatformWithMissingOptional()],
-      });
-      vi.mocked(backend.downloadAllFirmware).mockRejectedValue(new Error("io"));
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Download All"));
-        await Promise.resolve();
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      // CATCH-REJECTION assert: status string rendered
-      expect(container.textContent).toContain("Download failed: Error: io");
-    });
-  });
-
-  // ------------------------------------------------------------------
-  // K. BIOS tab — handleDownloadRequired
-  // ------------------------------------------------------------------
-  describe("bios tab — handleDownloadRequired", () => {
-    function biosPlatformWithMissingRequired(): FirmwarePlatformExt {
-      return makeBiosPlatform({
-        platform_slug: "snes",
-        files: [
-          {
-            id: 1,
-            file_name: "bios.rom",
-            size: 100,
-            md5: "x",
-            downloaded: false,
-            required: true,
-            description: "Required BIOS",
-            hash_valid: null,
-            classification: "required",
-          },
-        ],
-      });
-    }
-
-    it("calls downloadRequiredFirmware(slug) and refreshes BIOS on success", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [biosPlatformWithMissingRequired()],
-      });
-      vi.mocked(backend.downloadRequiredFirmware).mockResolvedValue({
-        success: true,
-        downloaded: 1,
-      });
-      const { getByText } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Download Required"));
-        await Promise.resolve();
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(vi.mocked(backend.downloadRequiredFirmware)).toHaveBeenCalledWith("snes");
-      expect(vi.mocked(backend.getFirmwareStatus)).toHaveBeenCalledTimes(2);
-    });
-
-    it("surfaces 'Download failed: <e>' when downloadRequiredFirmware throws", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [biosPlatformWithMissingRequired()],
-      });
-      vi.mocked(backend.downloadRequiredFirmware).mockRejectedValue(new Error("io"));
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Download Required"));
-        await Promise.resolve();
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      // CATCH-REJECTION assert: status string rendered
-      expect(container.textContent).toContain("Download failed: Error: io");
-    });
-
-    it("surfaces 'Download failed' fallback when result.success=false with no message", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [biosPlatformWithMissingRequired()],
-      });
-      vi.mocked(backend.downloadRequiredFirmware).mockResolvedValue({
-        success: false,
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Download Required"));
-        await Promise.resolve();
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("Download failed");
-    });
-  });
-
-  // ------------------------------------------------------------------
-  // L. BIOS tab — expand/collapse + hashIndicator + unknown summary
-  // ------------------------------------------------------------------
-  describe("bios tab — expand/collapse and file rendering", () => {
-    it("expands files on Show Files click and collapses on the same button", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "ok.bin",
-                size: 100,
-                md5: "x",
-                downloaded: true,
-                required: false,
-                description: "OK File",
-                hash_valid: true,
-                classification: "optional",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      // Initially collapsed — file name not rendered
-      expect(container.textContent).not.toContain("OK File");
-      await act(async () => {
-        fireEvent.click(getByText("Show Files (1)"));
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("OK File");
-      // Now collapse
-      await act(async () => {
-        fireEvent.click(getByText("Hide Files"));
-        await Promise.resolve();
-      });
-      expect(container.textContent).not.toContain("OK File");
-    });
-
-    it("renders hashIndicator ' ✓' for downloaded files with hash_valid=true", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "good.rom",
-                size: 100,
-                md5: "x",
-                downloaded: true,
-                required: false,
-                description: "Good",
-                hash_valid: true,
-                classification: "optional",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Show Files (1)"));
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("good.rom ✓");
-    });
-
-    it("renders hashIndicator ' ⚠' for downloaded files with hash_valid=false", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "bad.rom",
-                size: 100,
-                md5: "x",
-                downloaded: true,
-                required: false,
-                description: "Bad",
-                hash_valid: false,
-                classification: "optional",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Show Files (1)"));
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("bad.rom ⚠");
-    });
-
-    it("renders hashIndicator ' —' for downloaded files with hash_valid=null", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "unk.rom",
-                size: 100,
-                md5: "x",
-                downloaded: true,
-                required: false,
-                description: "Unk",
-                hash_valid: null,
-                classification: "optional",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Show Files (1)"));
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("unk.rom —");
-    });
-
-    it("renders a missing required file (red dot branch) when expanded", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "missing-req.rom",
-                size: 100,
-                md5: "x",
-                downloaded: false,
-                required: true,
-                description: "ReqMissing",
-                hash_valid: null,
-                classification: "required",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Show Files (1)"));
-        await Promise.resolve();
-      });
-      // Missing required → "Missing" suffix and red dot branch
-      expect(container.textContent).toContain("missing-req.rom — Missing");
-    });
-
-    it("renders a missing optional file (gray dot branch) when expanded", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "missing-opt.rom",
-                size: 100,
-                md5: "x",
-                downloaded: false,
-                required: false,
-                description: "OptMissing",
-                hash_valid: null,
-                classification: "optional",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Show Files (1)"));
-        await Promise.resolve();
-      });
-      // Missing optional → "Missing" suffix and gray dot branch
-      expect(container.textContent).toContain("missing-opt.rom — Missing");
-    });
-
-    it("renders the unrecognized-file footer when unknown files are present", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "mystery.rom",
-                size: 100,
-                md5: "x",
-                downloaded: true,
-                required: false,
-                description: "?",
-                hash_valid: null,
-                classification: "unknown",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        fireEvent.click(getByText("Show Files (1)"));
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("1 file(s) not recognized");
-    });
-  });
-
-  // ------------------------------------------------------------------
-  // M. BIOS tab — getBiosSummary indirect coverage via rendering
-  // ------------------------------------------------------------------
-  describe("bios tab — summary text", () => {
-    it("shows 'X / Y required' + 'All required ready' when all required are done and no optional missing", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "req.rom",
-                size: 100,
-                md5: "x",
-                downloaded: true,
-                required: true,
-                description: "Req",
-                hash_valid: true,
-                classification: "required",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("1 / 1 required");
-      expect(container.textContent).toContain("All required ready");
-    });
-
-    it("shows 'N optional missing' when all required are done but optional is missing", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "req.rom",
-                size: 100,
-                md5: "x",
-                downloaded: true,
-                required: true,
-                description: "Req",
-                hash_valid: true,
-                classification: "required",
-              },
-              {
-                id: 2,
-                file_name: "opt.rom",
-                size: 100,
-                md5: "x",
-                downloaded: false,
-                required: false,
-                description: "Opt",
-                hash_valid: null,
-                classification: "optional",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("All required ready (1 optional missing)");
-    });
-
-    it("shows 'N required missing — games may not launch' when required is incomplete", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "req.rom",
-                size: 100,
-                md5: "x",
-                downloaded: false,
-                required: true,
-                description: "Req",
-                hash_valid: null,
-                classification: "required",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("1 required missing");
-    });
-
-    it("falls back to 'X / Y files' summary when there are no required files", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "opt.rom",
-                size: 100,
-                md5: "x",
-                downloaded: true,
-                required: false,
-                description: "Opt",
-                hash_valid: true,
-                classification: "optional",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("1 / 1 files");
-      expect(container.textContent).toContain("All downloaded");
-    });
-
-    it("shows 'N missing' suffix when not all files are downloaded and no required files exist", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [
-              {
-                id: 1,
-                file_name: "opt.rom",
-                size: 100,
-                md5: "x",
-                downloaded: false,
-                required: false,
-                description: "Opt",
-                hash_valid: null,
-                classification: "optional",
-              },
-            ],
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(container.textContent).toContain("0 / 1 files");
-      expect(container.textContent).toContain("1 missing");
-    });
-  });
-
-  // ------------------------------------------------------------------
-  // N. BIOS tab — setSystemCore (core dropdown)
-  // ------------------------------------------------------------------
-  describe("bios tab — setSystemCore", () => {
-    it("does NOT render the dropdown when available_cores has <=1 entry", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [],
-            available_cores: [{ core_so: "snes9x.so", label: "snes9x", is_default: true }],
-          }),
-        ],
-      });
-      const { getByText } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(capturedDropdowns.length).toBe(0);
-    });
-
-    it("renders the core dropdown when available_cores has >1 entries", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [],
-            available_cores: [
-              { core_so: "snes9x.so", label: "snes9x", is_default: true },
-              { core_so: "mesen-s.so", label: "mesen-s", is_default: false },
-            ],
-            active_core_label: "snes9x",
-          }),
-        ],
-      });
-      const { getByText } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(capturedDropdowns.length).toBe(1);
-      const dropdown = capturedDropdowns[0]!;
-      expect(dropdown.label).toBe("Emulator Core");
-      expect(dropdown.rgOptions?.map((o) => o.data)).toEqual(["snes9x", "mesen-s"]);
-      expect(dropdown.rgOptions?.[0]?.label).toBe("snes9x (default)");
-      expect(dropdown.selectedOption).toBe("snes9x");
-    });
-
-    it("calls setSystemCore with empty label when default core is selected and dispatches romm_data_changed", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [],
-            available_cores: [
-              { core_so: "snes9x.so", label: "snes9x", is_default: true },
-              { core_so: "mesen-s.so", label: "mesen-s", is_default: false },
-            ],
-            active_core_label: "mesen-s",
-          }),
-        ],
-      });
-      const { getByText } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      const listener = vi.fn();
-      globalThis.addEventListener("romm_data_changed", listener);
-      try {
-        await act(async () => {
-          await capturedDropdowns[0]?.onChange?.({
-            data: "snes9x",
-            label: "snes9x (default)",
-          });
-        });
-        // Selecting the default core → label is "" sent to setSystemCore
-        expect(vi.mocked(backend.setSystemCore)).toHaveBeenCalledWith("snes", "");
-        expect(listener).toHaveBeenCalledTimes(1);
-        const ev = listener.mock.calls[0]?.[0] as CustomEvent;
-        expect(ev.detail).toEqual({
-          type: "core_changed",
-          platform_slug: "snes",
-        });
-      } finally {
-        globalThis.removeEventListener("romm_data_changed", listener);
-      }
-    });
-
-    it("calls setSystemCore with the explicit non-default label", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [],
-            available_cores: [
-              { core_so: "snes9x.so", label: "snes9x", is_default: true },
-              { core_so: "mesen-s.so", label: "mesen-s", is_default: false },
-            ],
-            active_core_label: "snes9x",
-          }),
-        ],
-      });
-      const { getByText } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      await act(async () => {
-        await capturedDropdowns[0]?.onChange?.({
-          data: "mesen-s",
-          label: "mesen-s",
-        });
-      });
-      expect(vi.mocked(backend.setSystemCore)).toHaveBeenCalledWith("snes", "mesen-s");
-    });
-
-    it("does NOT refresh BIOS or dispatch the event when setSystemCore returns success=false", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [],
-            available_cores: [
-              { core_so: "snes9x.so", label: "snes9x", is_default: true },
-              { core_so: "mesen-s.so", label: "mesen-s", is_default: false },
-            ],
-            active_core_label: "snes9x",
-          }),
-        ],
-      });
-      vi.mocked(backend.setSystemCore).mockResolvedValue({ success: false });
-      const { getByText } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      const listener = vi.fn();
-      globalThis.addEventListener("romm_data_changed", listener);
-      try {
-        await act(async () => {
-          await capturedDropdowns[0]?.onChange?.({
-            data: "mesen-s",
-            label: "mesen-s",
-          });
-        });
-        // refreshBios was called once on tab activation; not again on failure
-        expect(vi.mocked(backend.getFirmwareStatus)).toHaveBeenCalledTimes(1);
-        expect(listener).not.toHaveBeenCalled();
-      } finally {
-        globalThis.removeEventListener("romm_data_changed", listener);
-      }
-    });
-
-    it("falls back to default core label when active_core_label is absent", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [],
-            available_cores: [
-              { core_so: "snes9x.so", label: "snes9x", is_default: true },
-              { core_so: "mesen-s.so", label: "mesen-s", is_default: false },
-            ],
-            // No active_core_label
-          }),
-        ],
-      });
-      const { getByText } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      expect(capturedDropdowns[0]?.selectedOption).toBe("snes9x");
-    });
-
-    it("renders an inactive Core Field when active_core_label is set but only 1 available core exists", async () => {
-      vi.mocked(backend.getFirmwareStatus).mockResolvedValue({
-        success: true,
-        platforms: [
-          makeBiosPlatform({
-            platform_slug: "snes",
-            files: [],
-            available_cores: [{ core_so: "snes9x.so", label: "snes9x", is_default: true }],
-            active_core_label: "snes9x",
-          }),
-        ],
-      });
-      const { getByText, container } = render(<LibraryPage onBack={vi.fn()} />);
-      await flushAsync();
-      await act(async () => {
-        fireEvent.click(getByText("System"));
-        await Promise.resolve();
-        await Promise.resolve();
-      });
-      // No dropdown rendered, but the "Core" Field is.
-      expect(capturedDropdowns.length).toBe(0);
-      expect(container.textContent).toContain("snes9x");
     });
   });
