+++ src/components/LibraryPage.tsx
@@ -6,7 +6,6 @@ import {
   ToggleField,
   Spinner,
   DialogButton,
-  DropdownItem,
   Field,
   Focusable,
 } from "@decky/ui";
@@ -19,19 +18,8 @@ import {
   setAllCollectionsSync,
   saveCollectionPlatformGroups,
   getSettings,
-  getFirmwareStatus,
-  downloadAllFirmware,
-  downloadRequiredFirmware,
-  setSystemCore,
-  debugLog,
 } from "../api/backend";
-import type {
-  PlatformSyncSetting,
-  CollectionSyncSetting,
-  CollectionKind,
-  CollectionScope,
-  FirmwarePlatformExt,
-} from "../types";
+import type { PlatformSyncSetting, CollectionSyncSetting, CollectionKind, CollectionScope } from "../types";
 import { scrollToTop } from "../utils/scrollHelpers";
 import { detach } from "../utils/detach";
 
@@ -76,46 +64,12 @@ function favoritesDescription(romCount: number): string {
   return `Includes ${romCount} favorited games`;
 }
 
-function getBiosSummary(
-  requiredCount: number,
-  requiredDone: number,
-  allRequiredDone: boolean,
-  optionalMissing: number,
-  done: number,
-  total: number,
-  allDone: boolean,
-) {
-  if (requiredCount > 0 && allRequiredDone) {
-    return {
-      summaryLabel: `${requiredDone} / ${requiredCount} required`,
-      summaryDescription:
-        optionalMissing > 0 ? `All required ready (${optionalMissing} optional missing)` : "All required ready",
-    };
-  }
-  if (requiredCount > 0) {
-    return {
-      summaryLabel: `${requiredDone} / ${requiredCount} required`,
-      summaryDescription: `${requiredCount - requiredDone} required missing — games may not launch`,
-    };
-  }
-  return {
-    summaryLabel: `${done} / ${total} files`,
-    summaryDescription: allDone ? "All downloaded" : `${total - done} missing`,
-  };
-}
-
-function hashIndicator(hv: boolean | null): string {
-  if (hv === true) return " \u2713";
-  if (hv === false) return " \u26A0";
-  return " \u2014";
-}
-
 interface LibraryPageProps {
   onBack: () => void;
 }
 
 export const LibraryPage: FC<LibraryPageProps> = ({ onBack }) => {
-  const [activeTab, setActiveTab] = useState<"platforms" | "collections" | "system">("platforms");
+  const [activeTab, setActiveTab] = useState<"platforms" | "collections">("platforms");
 
   // --- Platforms tab state ---
   const [syncPlatforms, setSyncPlatforms] = useState<PlatformSyncSetting[]>([]);
@@ -146,16 +100,6 @@ export const LibraryPage: FC<LibraryPageProps> = ({ onBack }) => {
     }
     return favs[0] ?? null;
   }, [collections]);
-
-  // --- System tab state (per-system emulator core + BIOS files) ---
-  const [biosPlatforms, setBiosPlatforms] = useState<FirmwarePlatformExt[]>([]);
-  const [biosLoading, setBiosLoading] = useState(true);
-  const [biosError, setBiosError] = useState("");
-  const [serverOffline, setServerOffline] = useState(false);
-  const [downloading, setDownloading] = useState<string | null>(null);
-  const [biosStatus, setBiosStatus] = useState("");
-  const [expanded, setExpanded] = useState<Record<string, boolean>>({});
-  const systemLoaded = useRef(false);
 
   // Load sync platforms on mount
   useEffect(() => {
@@ -197,31 +141,6 @@ export const LibraryPage: FC<LibraryPageProps> = ({ onBack }) => {
     setActiveSubTab("my");
     setActiveTab("collections");
   };
-
-  async function refreshSystem() {
-    setBiosLoading(true);
-    setBiosError("");
-    try {
-      const result = await getFirmwareStatus();
-      if (result.success) {
-        setBiosPlatforms(result.platforms);
-        setServerOffline(result.server_offline ?? false);
-      } else {
-        setBiosError(result.message || "Failed to fetch firmware status");
-      }
-    } catch (e) {
-      setBiosError(`Failed to fetch firmware status: ${e}`);
-    }
-    setBiosLoading(false);
-  }
-
-  // Load System data (core + BIOS) lazily on first switch to the System tab
-  useEffect(() => {
-    if (activeTab === "system" && !systemLoaded.current) {
-      systemLoaded.current = true;
-      detach(refreshSystem());
-    }
-  }, [activeTab]);
 
   // --- Platforms tab handlers ---
   const handleToggle = async (id: number, enabled: boolean) => {
@@ -266,41 +185,6 @@ export const LibraryPage: FC<LibraryPageProps> = ({ onBack }) => {
     } catch {
       setCollections(previous);
     }
-  };
-
-  // --- System tab handlers (BIOS downloads) ---
-  const handleDownloadAll = async (platformSlug: string) => {
-    setDownloading(platformSlug);
-    setBiosStatus("");
-    try {
-      const result = await downloadAllFirmware(platformSlug);
-      if (result.success) {
-        setBiosStatus(result.message || `Downloaded ${result.downloaded} files`);
-        await refreshSystem();
-      } else {
-        setBiosStatus(result.message || "Download failed");
-      }
-    } catch (e) {
-      setBiosStatus(`Download failed: ${e}`);
-    }
-    setDownloading(null);
-  };
-
-  const handleDownloadRequired = async (platformSlug: string) => {
-    setDownloading(platformSlug);
-    setBiosStatus("");
-    try {
-      const result = await downloadRequiredFirmware(platformSlug);
-      if (result.success) {
-        setBiosStatus(result.message || `Downloaded ${result.downloaded} required files`);
-        await refreshSystem();
-      } else {
-        setBiosStatus(result.message || "Download failed");
-      }
-    } catch (e) {
-      setBiosStatus(`Download failed: ${e}`);
-    }
-    setDownloading(null);
   };
 
   // --- Platforms tab content ---
@@ -492,197 +376,6 @@ export const LibraryPage: FC<LibraryPageProps> = ({ onBack }) => {
     );
   };
 
-  // --- BIOS tab: platform rendering ---
-  const withGames = biosPlatforms.filter((p) => p.has_games);
-  const withoutGames = biosPlatforms.filter((p) => !p.has_games);
-
-  const renderBiosPlatform = (platform: FirmwarePlatformExt) => {
-    const total = platform.files.length;
-    const done = platform.files.filter((f) => f.downloaded).length;
-    const allDone = done === total;
-    const isDownloading = downloading === platform.platform_slug;
-    const isExpanded = expanded[platform.platform_slug] ?? false;
-
-    const requiredFiles = platform.files.filter((f) => f.classification === "required");
-    const unknownFiles = platform.files.filter((f) => f.classification === "unknown");
-    const requiredCount = requiredFiles.length;
-    const requiredDone = requiredFiles.filter((f) => f.downloaded).length;
-    const allRequiredDone = requiredDone === requiredCount;
-    const optionalMissing = platform.files.filter((f) => f.classification === "optional" && !f.downloaded).length;
-
-    const needsAttention = platform.has_games && !allRequiredDone;
-    const { summaryLabel, summaryDescription } = getBiosSummary(
-      requiredCount,
-      requiredDone,
-      allRequiredDone,
-      optionalMissing,
-      done,
-      total,
-      allDone,
-    );
-    const hasRequiredMissing = requiredCount > 0 && !allRequiredDone;
-    const hasOptionalMissing = optionalMissing > 0;
-
-    const hasMultipleCores = !!platform.available_cores && platform.available_cores.length > 1;
-
-    return (
-      <PanelSection
-        key={platform.platform_slug}
-        title={`${platform.platform_slug}${needsAttention ? " — BIOS needed" : ""}`}
-      >
-        {/* Emulator core selection is the primary per-system concern (#923),
-            shown above the BIOS file management. */}
-        {hasMultipleCores && (
-          <>
-            <PanelSectionRow>
-              <DropdownItem
-                label="Emulator Core"
-                rgOptions={[
-                  ...platform.available_cores!.map((c) => ({
-                    data: c.label,
-                    label: c.is_default ? `${c.label} (default)` : c.label,
-                  })),
-                ]}
-                selectedOption={
-                  platform.active_core_label || platform.available_cores!.find((c) => c.is_default)?.label || ""
-                }
-                onChange={(option: { data: string }) => {
-                  detach(
-                    (async () => {
-                      const defaultCore = platform.available_cores?.find((c) => c.is_default);
-                      const label = option.data === defaultCore?.label ? "" : option.data;
-                      detach(
-                        debugLog(
-                          `setSystemCore: slug=${platform.platform_slug} label=${label} (selected=${option.data})`,
-                        ),
-                      );
-                      try {
-                        const result = await setSystemCore(platform.platform_slug, label);
-                        detach(debugLog(`setSystemCore: result success=${result.success}`));
-                        if (result.success) {
-                          await refreshSystem();
-                          globalThis.dispatchEvent(
-                            new CustomEvent("romm_data_changed", {
-                              detail: { type: "core_changed", platform_slug: platform.platform_slug },
-                            }),
-                          );
-                        }
-                      } catch (e) {
-                        detach(debugLog(`setSystemCore: error: ${e}`));
-                      }
-                    })(),
-                  );
-                }}
-              />
-            </PanelSectionRow>
-            <PanelSectionRow>
-              <div style={{ fontSize: "11px", color: "#ffb74d", padding: "0 16px 4px" }}>
-                Switching cores may affect save compatibility
-              </div>
-            </PanelSectionRow>
-          </>
-        )}
-        {platform.active_core_label && !hasMultipleCores && (
-          <PanelSectionRow>
-            <Field label="Emulator Core" description={platform.active_core_label} />
-          </PanelSectionRow>
-        )}
-        <PanelSectionRow>
-          <Field label={summaryLabel} description={summaryDescription} />
-        </PanelSectionRow>
-        <PanelSectionRow>
-          <ButtonItem
-            layout="below"
-            onClick={() =>
-              setExpanded((prev) => ({
-                ...prev,
-                [platform.platform_slug]: !prev[platform.platform_slug],
-              }))
-            }
-          >
-            {isExpanded ? "Hide Files" : `Show Files (${total})`}
-          </ButtonItem>
-        </PanelSectionRow>
-        {isExpanded && (
-          <Focusable>
-            {platform.files.map((file) => {
-              let dotColor: string;
-              if (file.classification === "unknown") {
-                dotColor = "#d4a72c";
-              } else if (file.downloaded) {
-                dotColor = "#5ba32b";
-              } else if (file.classification === "required") {
-                dotColor = "#d94126";
-              } else {
-                dotColor = "#8f98a0";
-              }
-              return (
-                <PanelSectionRow key={file.id}>
-                  <Field
-                    label={
-                      <span style={{ display: "flex", alignItems: "center", gap: "6px" }}>
-                        <span
-                          style={{
-                            display: "inline-block",
-                            width: "8px",
-                            height: "8px",
-                            borderRadius: "50%",
-                            backgroundColor: dotColor,
-                            flexShrink: 0,
-                          }}
-                        />
-                        {`${file.description || file.file_name} (${file.classification})`}
-                      </span>
-                    }
-                    description={
-                      file.downloaded
-                        ? `${file.file_name}${hashIndicator(file.hash_valid)}`
-                        : `${file.file_name} — Missing`
-                    }
-                  />
-                </PanelSectionRow>
-              );
-            })}
-            {unknownFiles.length > 0 && (
-              <PanelSectionRow>
-                <Field
-                  label={`${unknownFiles.length} file(s) not recognized`}
-                  description="Report at github.com/danielcopper/decky-romm-sync/issues if needed."
-                />
-              </PanelSectionRow>
-            )}
-          </Focusable>
-        )}
-        {hasRequiredMissing && !serverOffline && (
-          <PanelSectionRow>
-            <ButtonItem
-              layout="below"
-              onClick={() => {
-                detach(handleDownloadRequired(platform.platform_slug));
-              }}
-              disabled={isDownloading}
-            >
-              {isDownloading ? "Downloading..." : "Download Required"}
-            </ButtonItem>
-          </PanelSectionRow>
-        )}
-        {!allDone && (hasOptionalMissing || hasRequiredMissing) && !serverOffline && (
-          <PanelSectionRow>
-            <ButtonItem
-              layout="below"
-              onClick={() => {
-                detach(handleDownloadAll(platform.platform_slug));
-              }}
-              disabled={isDownloading}
-            >
-              {isDownloading ? "Downloading..." : "Download All"}
-            </ButtonItem>
-          </PanelSectionRow>
-        )}
-      </PanelSection>
-    );
-  };
-
   // --- Render ---
   return (
     <>
@@ -723,70 +416,11 @@ export const LibraryPage: FC<LibraryPageProps> = ({ onBack }) => {
         >
           Collections
         </DialogButton>
-        <DialogButton
-          style={{
-            flex: 1,
-            minWidth: 0,
-            padding: "10px 0",
-            opacity: activeTab === "system" ? 1 : 0.5,
-            borderBottom: activeTab === "system" ? "2px solid #1a9fff" : "2px solid transparent",
-          }}
-          onClick={() => setActiveTab("system")}
-        >
-          System
-        </DialogButton>
       </Focusable>
 
       {activeTab === "platforms" && <PanelSection title="Platforms">{renderPlatformsContent()}</PanelSection>}
 
       {activeTab === "collections" && <>{renderCollectionsContent()}</>}
-
-      {activeTab === "system" && (
-        <>
-          <PanelSection title="System">
-            <PanelSectionRow>
-              <div style={{ fontSize: "11px", color: "#8f98a0", padding: "0 16px 4px" }}>
-                Per-system emulator core and BIOS files. The active core determines which BIOS files a system needs.
-              </div>
-            </PanelSectionRow>
-            {biosLoading && (
-              <PanelSectionRow>
-                <Field label="Loading firmware status..." />
-              </PanelSectionRow>
-            )}
-
-            {biosError && (
-              <PanelSectionRow>
-                <Field label="Error" description={biosError} />
-              </PanelSectionRow>
-            )}
-
-            {serverOffline && (
-              <PanelSectionRow>
-                <Field
-                  label="Server offline"
-                  description="RomM server is unreachable. Downloads unavailable, but core switching still works."
-                />
-              </PanelSectionRow>
-            )}
-
-            {!biosLoading && !biosError && biosPlatforms.length === 0 && (
-              <PanelSectionRow>
-                <Field label="No firmware files found" />
-              </PanelSectionRow>
-            )}
-
-            {biosStatus && (
-              <PanelSectionRow>
-                <Field label={biosStatus} />
-              </PanelSectionRow>
-            )}
-          </PanelSection>
-
-          {withGames.map(renderBiosPlatform)}
-          {withoutGames.map(renderBiosPlatform)}
-        </>
-      )}
     </>
   );
 };
